backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
web3==6.15.0
py-solc-x==2.0.2
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
facenet-pytorch==2.5.3
librosa==0.10.1
soundfile==0.12.1
